improve entrypoint
